+++ include/vectors.h
@@ -7,14 +7,14 @@
 #define VECTORS_H
 
 typedef struct {
-  double x;
-  double y;
+  float x;
+  float y;
 } vec2;
 
 typedef struct {
-  double x;
-  double y;
-  double z;
+  float x;
+  float y;
+  float z;
 } vec3;
 
 #endif  // VECTORS_H
